requirements.txt
gym==0.21.0
numpy~=1.23.4
setuptools~=60.2.0
rich~=12.6.0
pygame~=2.1.2
pillow~=9.0.1
stable-baselines3~=1.6.2
torch~=1.13.1
tensorboard~=2.11.0